+++ flake.nix
@@ -10,11 +10,12 @@
     nixpkgs-22_05.url = "github:NixOS/nixpkgs?ref=nixos-22.05";
     nixpkgs-21_11.url = "github:NixOS/nixpkgs?ref=nixos-21.11";
     pre-commit-hooks.url = "github:cachix/pre-commit-hooks.nix";
+    horizon-core.url = "git+https://gitlab.horizon-haskell.net/package-sets/horizon-core";
     autodocodec.url = "github:NorfairKing/autodocodec";
     autodocodec.flake = false;
     safe-coloured-text.url = "github:NorfairKing/safe-coloured-text";
     safe-coloured-text.flake = false;
-    sydtest.url = "github:NorfairKing/sydtest";
+    sydtest.url = "github:NorfairKing/sydtest?ref=forward-compatibility";
     sydtest.flake = false;
   };
 
@@ -25,26 +26,40 @@
     , nixpkgs-22_05
     , nixpkgs-21_11
     , pre-commit-hooks
+    , horizon-core
     , autodocodec
     , safe-coloured-text
     , sydtest
     }:
     let
       system = "x86_64-linux";
+      overlays = [
+        self.overlays.${system}
+        (import (autodocodec + "/nix/overlay.nix"))
+        (import (safe-coloured-text + "/nix/overlay.nix"))
+        (import (sydtest + "/nix/overlay.nix"))
+      ];
       pkgsFor = nixpkgs: import nixpkgs {
         inherit system;
+        inherit overlays;
+      };
+      horizonPkgs = import nixpkgs {
+        inherit system;
         overlays = [
-          self.overlays.${system}
-          (import (autodocodec + "/nix/overlay.nix"))
-          (import (safe-coloured-text + "/nix/overlay.nix"))
-          (import (sydtest + "/nix/overlay.nix"))
-        ];
+          (final: prev: {
+            haskellPackages = prev.haskellPackages.override (old: {
+              overrides = final.lib.composeExtensions (old.overrides or (_: _: { })) (self: super:
+                horizon-core.legacyPackages.${system} // super
+              );
+            });
+          })
+        ] ++ overlays;
       };
       pkgs = pkgsFor nixpkgs;
     in
     {
       overlays.${system} = import ./nix/overlay.nix;
-      packages.${system}.default = pkgs.haskellPackages.validityRelease;
+      packages.${system} = pkgs.haskellPackages.validityPackages;
       checks.${system} =
         let
           backwardCompatibilityCheckFor = nixpkgs:
@@ -59,6 +74,8 @@
           backwardCompatibilityChecks = pkgs.lib.mapAttrs (_: nixpkgs: backwardCompatibilityCheckFor nixpkgs) allNixpkgs;
         in
         backwardCompatibilityChecks // {
+          forwardCompatibility = horizonPkgs.haskellPackages.validityRelease;
+          release = pkgs.haskelLPackages.validityRelease;
           pre-commit = pre-commit-hooks.lib.${system}.run {
             src = ./.;
             hooks = {