+++ genvalidity-hspec-hashable/src/Test/Validity/Hashable.hs
@@ -2,6 +2,7 @@
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
+{-# OPTIONS_GHC -Wno-redundant-constraints #-}
 
 -- | Hashable properties
 --
@@ -71,19 +72,19 @@ checkGen gen genname s =
   parallel $ do
     let name = nameOf @a
         hashablestr = unwords ["hashWithSalt :: Int ->", name, "-> Int"]
-    describe ("Hashable " ++ name) $
-      describe hashablestr $
-        it
-          ( unwords
-              [ "satisfies (a == b) => (hashWithSalt n a) ==",
-                "(hashWithSalt n b), for every n and for",
-                genname,
-                name
-              ]
-          )
-          $ let ss (a, b) = (,) <$> s a <*> s b
-             in forAllShrink gen ss $ \(a1, a2) ->
-                  forAllValid $ \int ->
-                    when (a1 == a2) $
-                      let h = hashWithSalt int
-                       in h a1 `shouldBe` h a2
+    describe ("Hashable " ++ name)
+      $ describe hashablestr
+      $ it
+        ( unwords
+            [ "satisfies (a == b) => (hashWithSalt n a) ==",
+              "(hashWithSalt n b), for every n and for",
+              genname,
+              name
+            ]
+        )
+      $ let ss (a, b) = (,) <$> s a <*> s b
+         in forAllShrink gen ss $ \(a1, a2) ->
+              forAllValid $ \int ->
+                when (a1 == a2) $
+                  let h = hashWithSalt int
+                   in h a1 `shouldBe` h a2