+++ genvalidity-unordered-containers/default.nix
@@ -4,7 +4,7 @@
 }:
 mkDerivation {
   pname = "genvalidity-unordered-containers";
-  version = "1.0.0.0";
+  version = "1.0.0.1";
   src = ./.;
   libraryHaskellDepends = [
     base genvalidity hashable QuickCheck unordered-containers validity