+++ genvalidity-unordered-containers/src/Data/GenValidity/HashSet.hs
@@ -1,4 +1,5 @@
-{-# OPTIONS_GHC -fno-warn-orphans #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
+{-# OPTIONS_GHC -Wno-redundant-constraints #-}
 
 module Data.GenValidity.HashSet where