+++ genvalidity-hspec-hashable/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.0.0.1] - 2023-10-04
+
+* Compatibility with `hashable >= 1.4`
+
 ## [1.0.0.0] - 2021-11-20
 
 * Compatibility with `genvalidity >= 1.0.0.0`