+++ genvalidity-persistent/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.0.0.2] - 2023-10-04
+
+### Changed
+
+* Forward compatibility with `persistent` requiring `TypeOperators`.
+
 ## [1.0.0.1] - 2022-08-30
 
 ### Added